Implemented slice and string operations
Moved memory operations to internal runtime package.
Implemented builtins for handling string and slice operations.
Refactored value handling to track which values on the heap.
Debug info about values are now stored with the value.

+++ src/runtime/type.go
@@ -44,7 +44,7 @@ const (
 
 type typeDescriptor struct {
 	name     *string
-	size     int64
+	size     uintptr
 	kind     BasicKind
 	methods  *methodTable
 	fields   *fieldTable
